requirements.txt
python-dotenv
langchain-google-genai